set default to use Double precision

+++ cbits/type.h
@@ -6,10 +6,10 @@ typedef float F;
 typedef double F;
 #else
 #ifdef FLYCHECK
-typedef float F;
+typedef double F;
 #else
 COMPILATION ERROR: You need to either provide the pre-processor direvtive USE_FLOAT or USE_DOUBLE! See the package.yaml file of grenade!;
-typedef float F;                /* To prevent other errors! */
+typedef double F;                /* To prevent other errors! */
 #endif
 #endif
 #endif
